+++ Src/CSharpier.VSCode/package.json
@@ -1,7 +1,7 @@
 {
     "name": "csharpier-vscode",
     "license": "MIT",
-    "version": "0.1.0",
+    "version": "1.0.0",
     "engines": {
         "vscode": "^1.60.0"
     },
